@@ -1,15 +1,20 @@
 package com.sidpatchy.clairebot.Listener;
 
-import com.sidpatchy.clairebot.Embed.Commands.Regular.AvatarEmbed;
+import com.sidpatchy.clairebot.API.APIUser;
 import com.sidpatchy.clairebot.Main;
+import com.sidpatchy.clairebot.Util.Leveling.LevelingTools;
 import org.javacord.api.entity.message.Message;
+import org.javacord.api.entity.message.MessageAuthor;
 import org.javacord.api.entity.message.MessageBuilder;
 import org.javacord.api.entity.message.mention.AllowedMentionsBuilder;
+import org.javacord.api.entity.server.Server;
 import org.javacord.api.event.message.MessageCreateEvent;
 import org.javacord.api.listener.message.MessageCreateListener;
 
+import java.io.IOException;
 import java.util.List;
 import java.util.Random;
+import java.util.random.RandomGenerator;
 import java.util.regex.Pattern;
 
 public class MessageCreate implements MessageCreateListener {
@@ -17,6 +22,13 @@ public class MessageCreate implements MessageCreateListener {
     public void onMessageCreate(MessageCreateEvent event) {
         Message message = event.getMessage();
         String messageContent = message.getContent();
+        Server server = message.getServer().orElse(null);
+        MessageAuthor messageAuthor = message.getAuthor();
+        APIUser apiUser = new APIUser(messageAuthor.getIdAsString());
+
+        if (messageAuthor.isBotUser() || messageAuthor.isYourself()) {
+            return;
+        }
 
         // ClaireBot on top!!
         List<String> onTopResponses = Main.getClaireBotOnTopResponses();
@@ -34,6 +46,20 @@ public class MessageCreate implements MessageCreateListener {
                         .send(message.getChannel());
 
                 break;
+            }
+        }
+
+        // Grant between 0 and 8 points
+        if (server != null) {
+            Integer currentPoints = LevelingTools.getUserPoints(messageAuthor.getIdAsString(), "global");
+            RandomGenerator randomGenerator = RandomGenerator.getDefault();
+            Integer pointsToGrant = randomGenerator.nextInt(8);
+            try {
+                apiUser.updateUserPointsGuildID(server.getIdAsString(), pointsToGrant);
+                apiUser.updateUserPointsGuildID("global", pointsToGrant);
+
+            } catch (IOException e) {
+                throw new RuntimeException(e);
             }
         }
     }
